+skeleton

+++ src/components/product/ProductCard.vue
@@ -7,7 +7,7 @@ import type { IProduct } from '../../models/IProduct'
 import { useUserStore } from '@/stores/user'
 
 const props = defineProps<{
-  list: IProduct
+  list: IProduct | null
 }>()
 const { list } = props
 const url = import.meta.env.VITE_BASE_URL_BD
@@ -17,23 +17,33 @@ const userStore = useUserStore()
 const { state } = storeToRefs(userStore)
 
 const clickHandler = async () => {
-  await cartStore.addOne(state.value.user.id, list._id)
+  if (list) {
+    await cartStore.addOne(state.value.user.id, list._id)
+  }
 }
 const removeHandler = async () => {
-  await cartStore.deleteOne(state.value.user.id, list._id)
+  if (list) {
+    await cartStore.deleteOne(state.value.user.id, list._id)
+  }
 }
 </script>
 
 <template>
-  <li class="product">
+  <li v-if="list" class="product">
     <RouterLink class="productLink" :to="{ name: 'product', params: { id: list._id } }">
-      <img class="productImg" width="200px" :src="url + '/' + list.img" :alt="list.title" />
+      <img
+        class="productImg"
+        width="200px"
+        height="200px"
+        :src="url + '/' + list.img"
+        :alt="list.title"
+      />
       <h3 class="productTitle">{{ list.title }}</h3>
     </RouterLink>
     <div class="productInfo">
       <div class="productContainer">
         <div class="productPriceLabel">Цена:</div>
-        <p class="productPrice">{{ list.price }}</p>
+        <p class="productPrice">{{ list.price }}MDL</p>
       </div>
       <div class="productControls">
         <div class="outOfStockContainer" v-if="cartStore.isInCart(list._id)">
@@ -52,9 +62,26 @@ const removeHandler = async () => {
       </div>
     </div>
   </li>
+  <li v-else class="product">
+    <div class="productLink">
+      <div class="skeletonImg skeleton"></div>
+      <h3 class="productTitle skeleton"></h3>
+    </div>
+    <div class="productInfo">
+      <div class="productContainer">
+        <div class="productPriceLabel skeleton">тест:</div>
+        <p class="productPrice skeleton">000000</p>
+      </div>
+      <div class="productControls">
+        <div class="outOfStockContainer skeleton">
+          <div class="skeletonBtn skeleton"></div>
+        </div>
+      </div>
+    </div>
+  </li>
 </template>
 
-<style scoped>
+<style scoped lang="scss">
 .product {
   width: 300px;
   padding: 20px 30px;
@@ -62,41 +89,47 @@ const removeHandler = async () => {
   flex-direction: column;
   align-items: center;
   justify-content: space-between;
-  background-color: var(--white);
+  background-color: var(--tertiary);
   border-radius: 20px;
-  box-shadow: inset 0 0 10px var(--white);
 
   box-shadow:
-    rgba(240, 46, 170, 0.4) 0px 5px,
-    rgba(240, 46, 170, 0.3) 0px 10px,
-    rgba(240, 46, 170, 0.2) 0px 15px,
-    rgba(240, 46, 170, 0.1) 0px 20px,
-    rgba(240, 46, 170, 0.05) 0px 25px;
+    rgba(187, 134, 252, 0.4) 0px 5px,
+    rgba(187, 134, 252, 0.3) 0px 10px,
+    rgba(187, 134, 252, 0.2) 0px 15px,
+    rgba(187, 134, 252, 0.1) 0px 20px,
+    rgba(187, 134, 252, 0.05) 0px 25px;
 
   transition: all 0.3s ease-in-out;
-}
-.product:hover,
-.product:focus {
-  box-shadow:
-    rgba(240, 46, 170, 0.4) -5px 5px,
-    rgba(240, 46, 170, 0.3) -10px 10px,
-    rgba(240, 46, 170, 0.2) -15px 15px,
-    rgba(240, 46, 170, 0.1) -20px 20px,
-    rgba(240, 46, 170, 0.05) -25px 25px;
+  &:hover,
+  &:focus {
+    box-shadow:
+      rgba(3, 218, 197, 0.4) -5px 5px,
+      rgba(3, 218, 197, 0.3) -10px 10px,
+      rgba(3, 218, 197, 0.2) -15px 15px,
+      rgba(3, 218, 197, 0.1) -20px 20px,
+      rgba(3, 218, 197, 0.05) -25px 25px;
+  }
 }
 .productLink {
   display: block;
+  width: 100%;
 
   text-decoration: none;
-  color: var(--black);
+  color: var(--surface);
 }
 .productTitle {
   margin-top: 15px;
+
+  height: 2lh;
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+  text-overflow: ellipsis;
+
   font-size: 20px;
   font-weight: 700;
   line-height: 24px;
-
-  text-align: justify;
 }
 
 .productImg {
@@ -107,11 +140,15 @@ const removeHandler = async () => {
 .productInfo {
   margin-top: 10px;
   width: 100%;
+  height: 40px;
   display: flex;
   justify-content: space-between;
   align-items: center;
 }
 .productContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
 }
 .productPriceLabel {
   margin-right: 15px;
@@ -142,5 +179,36 @@ const removeHandler = async () => {
   display: flex;
   align-self: center;
   color: var(--negative);
+}
+
+/* skeleton */
+.skeleton {
+  width: 100%;
+  background-color: var(--on-white-skeleton);
+  color: var(--on-white-skeleton);
+  cursor: none;
+  border-radius: 20px;
+  background: linear-gradient(
+      100deg,
+      rgba(255, 255, 255, 0) 40%,
+      rgba(255, 255, 255, 0.5) 50%,
+      rgba(255, 255, 255, 0) 60%
+    )
+    var(--on-white-skeleton);
+  background-size: 200% 100%;
+  background-position-x: 180%;
+  animation: 1s loading ease-in-out infinite;
+}
+.skeletonBtn {
+  width: 40px;
+  height: 40px;
+  background-color: var(--on-white-skeleton);
+  border-radius: 5px;
+}
+.skeletonImg {
+  width: 240px;
+  height: 200px;
+  background-color: var(--on-white-skeleton);
+  border-radius: 20px;
 }
 </style>
